refactor(backend): Cleanup auth related stuff
