+++ backend/routers/auth_router.py
@@ -1,36 +1,16 @@
 from typing import List
 
 import structlog
-from fastapi import APIRouter, Depends, HTTPException, status
-from fastapi.security import OAuth2PasswordBearer
+from fastapi import APIRouter, Depends
 from sqlalchemy.orm import Session
 
 from backend import models, schemas
 from backend.database import get_db
-from backend.supabase import supabase
+from backend.utils.auth import get_current_user
 
 logger = structlog.get_logger()
 
 auth_router = APIRouter()
-
-oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
-
-
-def get_current_user(token: str = Depends(oauth2_scheme)):
-    try:
-        user_info = supabase.auth.get_user(token)
-        if user_info is None:
-            raise HTTPException(
-                status_code=status.HTTP_401_UNAUTHORIZED,
-                detail="Invalid authentication credentials"
-            )
-        return user_info
-
-    except Exception as e:
-        raise HTTPException(
-            status_code=status.HTTP_400_BAD_REQUEST,
-            detail=f"{e}"
-        )
 
 
 @auth_router.get("/get_user", description="Returns Supabase user information")
