+++ backend/services/supabase.py
@@ -1,4 +1,5 @@
-from backend.settings import settings
 from supabase import Client, create_client
+
+from backend.settings import settings
 
 supabase: Client = create_client(settings.supabase_url, settings.supabase_key)
